+++ src/main/java/refinedstorage/block/BlockCrafter.java
@@ -32,7 +32,12 @@ public class BlockCrafter extends BlockNode {
     }
 
     @Override
-    public boolean hasOppositeFacingOnSneakPlace() {
+    public boolean canBePlacedOpposite() {
         return true;
+    }
+
+    @Override
+    public EnumDirectionType getDirectionType() {
+        return EnumDirectionType.ANY;
     }
 }