+++ src/main/java/refinedstorage/block/BlockExternalStorage.java
@@ -46,7 +46,12 @@ public class BlockExternalStorage extends BlockNode {
     }
 
     @Override
-    public boolean hasOppositeFacingOnSneakPlace() {
+    public boolean canBePlacedOpposite() {
         return true;
+    }
+
+    @Override
+    public EnumDirectionType getDirectionType() {
+        return EnumDirectionType.ANY;
     }
 }