+++ src/main/java/refinedstorage/block/BlockQuartzEnrichedIron.java
@@ -4,4 +4,9 @@ public class BlockQuartzEnrichedIron extends BlockBase {
     public BlockQuartzEnrichedIron() {
         super("quartz_enriched_iron_block");
     }
+
+    @Override
+    public EnumDirectionType getDirectionType() {
+        return null;
+    }
 }