+++ src/main/java/refinedstorage/block/BlockSolderer.java
@@ -1,6 +1,5 @@
 package refinedstorage.block;
 
-import net.minecraft.block.properties.IProperty;
 import net.minecraft.block.properties.PropertyBool;
 import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
@@ -39,11 +38,9 @@ public class BlockSolderer extends BlockNode {
 
     @Override
     protected BlockStateContainer createBlockState() {
-        return new BlockStateContainer(this, new IProperty[]{
-            DIRECTION,
-            CONNECTED,
-            WORKING
-        });
+        return createBlockStateBuilder()
+            .add(WORKING)
+            .build();
     }
 
     @Override