+++ src/main/java/refinedstorage/block/BlockStorage.java
@@ -1,6 +1,5 @@
 package refinedstorage.block;
 
-import net.minecraft.block.properties.IProperty;
 import net.minecraft.block.properties.PropertyEnum;
 import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
@@ -41,11 +40,9 @@ public class BlockStorage extends BlockNode {
 
     @Override
     protected BlockStateContainer createBlockState() {
-        return new BlockStateContainer(this, new IProperty[]{
-            DIRECTION,
-            CONNECTED,
-            TYPE
-        });
+        return createBlockStateBuilder()
+            .add(TYPE)
+            .build();
     }
 
     @Override
@@ -106,5 +103,10 @@ public class BlockStorage extends BlockNode {
     @Override
     public Item createItemForBlock() {
         return new ItemBlockStorage();
+    }
+
+    @Override
+    public EnumDirectionType getDirectionType() {
+        return null;
     }
 }